+++ BitFlipProb.h
@@ -1,14 +1,14 @@
+#include "Mutator.h"
 #ifndef BITFLIPPROB_H
 #define BITFLIPPROB_H
 
-#include "Mutator.h"
 
 class BitFlipProb : public Mutator{
     private:
 
     public:
     BitFlipProb();
-    Individual mutate(Individual I, int p);
+    Individual mutate(Individual I, double p);
     ~BitFlipProb();
 
 };
